cland tidy fixes

+++ tasks/gutyansky_a_matrix_column_sum/mpi/src/ops_mpi.cpp
@@ -5,6 +5,7 @@
 #include <cstddef>
 #include <cstdint>
 #include <vector>
+#include <utility>
 
 #include "gutyansky_a_matrix_column_sum/common/include/common.hpp"
 
@@ -68,10 +69,10 @@ bool GutyanskyAMatrixColumnSumMPI::RunImpl() {
     }
 
     MPI_Scatterv(GetInput().data.data(), send_counts.data(), locs.data(), MPI_INTEGER4, input_data_chunk.data(),
-                 elements_count, MPI_INTEGER4, 0, MPI_COMM_WORLD);
+                 static_cast<int>(elements_count), MPI_INTEGER4, 0, MPI_COMM_WORLD);
   } else {
-    MPI_Scatterv(nullptr, nullptr, nullptr, MPI_INTEGER4, input_data_chunk.data(), elements_count, MPI_INTEGER4, 0,
-                 MPI_COMM_WORLD);
+    MPI_Scatterv(nullptr, nullptr, nullptr, MPI_INTEGER4, input_data_chunk.data(),
+                 static_cast<int>(elements_count), MPI_INTEGER4, 0, MPI_COMM_WORLD);
   }
 
   for (size_t i = 0; i < rows_count; i++) {
